requirements.txt
django
djangorestframework
pillow
twilio
django-environ
django-cors-headers
dotenv
python-dotenv